test(ab): consolidate MCP UI AB coverage
